feat: command line model generator

+++ subprojects/store-dse/src/main/java/tools/refinery/store/dse/strategy/BestFirstStoreManager.java
@@ -79,7 +79,7 @@ public class BestFirstStoreManager {
 		startExploration(initial, 1);
 	}
 
-	public void startExploration(Version initial, int randomSeed) {
+	public void startExploration(Version initial, long randomSeed) {
 		BestFirstExplorer bestFirstExplorer = new BestFirstExplorer(this, modelStore.createModelForState(initial),
 				randomSeed);
 		bestFirstExplorer.explore();
